🔧 Chore: domain 패키지 정리
- Model 파일 이름 규칙 적용
- group 폴더 삭제
- usecase 도메인 별 파일 정리

Related to: #334
