+++ app/src/main/java/com/mongmong/namo/data/remote/TermApiService.kt
@@ -1,6 +1,6 @@
 package com.mongmong.namo.data.remote
 
-import com.mongmong.namo.domain.model.TermBody
+import com.mongmong.namo.data.dto.TermBody
 import com.mongmong.namo.presentation.config.BaseResponse
 import retrofit2.http.Body
 import retrofit2.http.POST
