+++ app/src/main/java/com/mongmong/namo/domain/repositories/TermRepository.kt
@@ -1,6 +1,6 @@
 package com.mongmong.namo.domain.repositories
 
-import com.mongmong.namo.domain.model.TermBody
+import com.mongmong.namo.data.dto.TermBody
 
 interface TermRepository {
     suspend fun postTerms(
