--- app/src/main/java/com/mongmong/namo/domain/usecases/DeleteMoimDiaryUseCase.kt
@@ -1,4 +1,0 @@
-package com.mongmong.namo.domain.usecases
-
-class DeleteMoimDiaryUseCase {
-}
