+++ app/src/main/java/com/mongmong/namo/domain/usecases/category/FindCategoryUseCase.kt
@@ -1,4 +1,4 @@
-package com.mongmong.namo.domain.usecases
+package com.mongmong.namo.domain.usecases.category
 
 import com.mongmong.namo.domain.model.Category
 import com.mongmong.namo.domain.repositories.CategoryRepository
