+++ app/src/main/java/com/mongmong/namo/domain/usecases/diary/DeleteMoimDiaryUseCase.kt
@@ -1,0 +1,4 @@
+package com.mongmong.namo.domain.usecases.diary
+
+class DeleteMoimDiaryUseCase {
+}
