+++ app/src/main/java/com/mongmong/namo/domain/usecases/friend/DenyFriendRequestUseCase.kt
@@ -1,13 +1,13 @@
-package com.mongmong.namo.domain.usecases
+package com.mongmong.namo.domain.usecases.friend
 
 import com.mongmong.namo.data.dto.FriendBaseResponse
 import com.mongmong.namo.domain.repositories.FriendRepository
 import javax.inject.Inject
 
-class AcceptFriendRequestUseCase @Inject constructor(private val friendRepository: FriendRepository) {
+class DenyFriendRequestUseCase @Inject constructor(private val friendRepository: FriendRepository) {
     suspend fun execute(
         friendRequestId: Long
     ): FriendBaseResponse {
-        return friendRepository.acceptFriendRequest(friendRequestId)
+        return friendRepository.rejectFriendRequest(friendRequestId)
     }
 }
