+++ app/src/main/java/com/mongmong/namo/domain/usecases/image/UploadImageToS3UseCase.kt
@@ -1,4 +1,4 @@
-package com.mongmong.namo.domain.usecases
+package com.mongmong.namo.domain.usecases.image
 
 import android.net.Uri
 import com.mongmong.namo.domain.repositories.ImageRepository
