+++ app/src/main/java/com/mongmong/namo/presentation/di/RepositoryModule.kt
@@ -4,27 +4,27 @@ import android.content.Context
 import com.mongmong.namo.data.datasource.auth.RemoteAuthDataSource
 import com.mongmong.namo.data.datasource.category.RemoteCategoryDataSource
 import com.mongmong.namo.data.datasource.diary.ActivityDataSource
-import com.mongmong.namo.data.datasource.schedule.RemoteScheduleDataSource
 import com.mongmong.namo.data.datasource.diary.RemoteDiaryDataSource
-import com.mongmong.namo.data.datasource.group.GroupDataSource
+import com.mongmong.namo.data.datasource.friend.RemoteFriendDataSource
 import com.mongmong.namo.data.datasource.s3.ImageDataSource
+import com.mongmong.namo.data.datasource.schedule.RemoteScheduleDataSource
 import com.mongmong.namo.data.datasource.terms.RemoteTermDataSource
-import com.mongmong.namo.data.remote.NetworkChecker
 import com.mongmong.namo.data.remote.DiaryApiService
+import com.mongmong.namo.data.remote.NetworkChecker
 import com.mongmong.namo.data.repositoriyImpl.ActivityRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.AuthRepositoryImpl
-import com.mongmong.namo.data.repositoriyImpl.ImageRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.CategoryRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.DiaryRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.FriendRepositoryImpl
+import com.mongmong.namo.data.repositoriyImpl.ImageRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.ScheduleRepositoryImpl
 import com.mongmong.namo.data.repositoriyImpl.TermRepositoryImpl
 import com.mongmong.namo.domain.repositories.ActivityRepository
 import com.mongmong.namo.domain.repositories.AuthRepository
-import com.mongmong.namo.domain.repositories.ImageRepository
 import com.mongmong.namo.domain.repositories.CategoryRepository
 import com.mongmong.namo.domain.repositories.DiaryRepository
 import com.mongmong.namo.domain.repositories.FriendRepository
+import com.mongmong.namo.domain.repositories.ImageRepository
 import com.mongmong.namo.domain.repositories.ScheduleRepository
 import com.mongmong.namo.domain.repositories.TermRepository
 import dagger.Module
