+++ app/src/main/java/com/mongmong/namo/presentation/di/UseCaseModule.kt
@@ -5,14 +5,14 @@ import com.mongmong.namo.domain.repositories.CategoryRepository
 import com.mongmong.namo.domain.repositories.DiaryRepository
 import com.mongmong.namo.domain.repositories.FriendRepository
 import com.mongmong.namo.domain.repositories.ImageRepository
-import com.mongmong.namo.domain.usecases.AcceptFriendRequestUseCase
-import com.mongmong.namo.domain.usecases.AddMoimDiaryUseCase
-import com.mongmong.namo.domain.usecases.DenyFriendRequestUseCase
-import com.mongmong.namo.domain.usecases.FindCategoryUseCase
-import com.mongmong.namo.domain.usecases.GetCategoriesUseCase
-import com.mongmong.namo.domain.usecases.GetActivitiesUseCase
-import com.mongmong.namo.domain.usecases.GetFriendsUseCase
-import com.mongmong.namo.domain.usecases.UploadImageToS3UseCase
+import com.mongmong.namo.domain.usecases.friend.AcceptFriendRequestUseCase
+import com.mongmong.namo.domain.usecases.diary.AddMoimDiaryUseCase
+import com.mongmong.namo.domain.usecases.friend.DenyFriendRequestUseCase
+import com.mongmong.namo.domain.usecases.category.FindCategoryUseCase
+import com.mongmong.namo.domain.usecases.category.GetCategoriesUseCase
+import com.mongmong.namo.domain.usecases.activity.GetActivitiesUseCase
+import com.mongmong.namo.domain.usecases.friend.GetFriendsUseCase
+import com.mongmong.namo.domain.usecases.image.UploadImageToS3UseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
