+++ app/src/main/java/com/mongmong/namo/presentation/ui/community/friend/FriendViewModel.kt
@@ -6,9 +6,9 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.mongmong.namo.domain.model.Friend
 import com.mongmong.namo.domain.repositories.FriendRepository
-import com.mongmong.namo.domain.usecases.AcceptFriendRequestUseCase
-import com.mongmong.namo.domain.usecases.DenyFriendRequestUseCase
-import com.mongmong.namo.domain.usecases.GetFriendsUseCase
+import com.mongmong.namo.domain.usecases.friend.AcceptFriendRequestUseCase
+import com.mongmong.namo.domain.usecases.friend.DenyFriendRequestUseCase
+import com.mongmong.namo.domain.usecases.friend.GetFriendsUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
