+++ app/src/main/java/com/mongmong/namo/presentation/ui/login/TermsFragment.kt
@@ -4,17 +4,11 @@ import android.content.Intent
 import android.content.res.ColorStateList
 import android.graphics.Color
 import android.net.Uri
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
 import android.widget.CompoundButton
-import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import com.mongmong.namo.R
 import com.mongmong.namo.databinding.FragmentTermsBinding
-import com.mongmong.namo.domain.model.TermBody
+import com.mongmong.namo.data.dto.TermBody
 import com.mongmong.namo.presentation.config.BaseFragment
 import com.mongmong.namo.presentation.config.Constants
 import com.mongmong.namo.presentation.ui.MainActivity
