+++ app/src/main/java/com/mongmong/namo/presentation/utils/StringConverter.kt
@@ -1,11 +1,11 @@
 package com.mongmong.namo.presentation.utils
 
-import com.mongmong.namo.domain.model.group.GroupMember
+import com.mongmong.namo.domain.model.Participant
 
 object StringConverter {
     @JvmStatic
-    fun getMembersText(memberList: List<GroupMember>): String {
+    fun getMembersText(memberList: List<Participant>): String {
         if (memberList.isEmpty()) return "없음"
-        return memberList.joinToString(", ") { it.userName }
+        return memberList.joinToString(", ") { it.nickname }
     }
 }
